chore: include README in NPM package

+++ scripts/release-npm.sh
@@ -1,13 +1,36 @@
 #!/bin/bash
 
+set -e
+
+usage() {
+    echo "$0 --tag=v0.0.0"
+}
+
+for i in "$@"; do
+    case $i in
+    --tag=*)
+        TAG="${i#*=}"
+        shift
+        ;;
+    *)
+        usage
+        exit 1
+        ;;
+    esac
+done
+
+if [ -z "$TAG" ]; then
+    usage
+    exit 1
+fi
+
 cd $(dirname $0)/..
 
 rm -rf dist
 
-VERSION="$1"
-export VERSION="${VERSION}"
+export VERSION="${TAG}"
 
-echo "Updating NPM package version to ${VERSION}"
+echo "Updating NPM package version to ${TAG}"
 
 mkdir -p dist/
 
@@ -18,4 +41,5 @@ done
 GOOS=windows GOARCH=amd64 go build -a -o dist/snyk-iac-rules.exe .
 
 cp packaging/npm/passthrough.js dist/snyk-iac-rules
+cp README.md dist/README.md
 envsubst < packaging/npm/package.json.in > dist/package.json
